create user moved to conftest, test names now is normal
