+++ app/main.py
@@ -1,4 +1,6 @@
 import dotenv
+
+
 dotenv.load_dotenv()
 
 
@@ -7,6 +9,7 @@ from fastapi import FastAPI
 
 from app.routers import status, users
 from app.database.engine import create_db_and_tables
+
 
 app = FastAPI()
 app.include_router(status.router)
