+++ tests/conftest.py
@@ -3,6 +3,13 @@ import dotenv
 import pytest
 import requests
 import json
+from faker import Faker
+from random import randint
+from http import HTTPStatus
+
+
+fake = Faker()
+
 
 @pytest.fixture(autouse=True, scope="session")
 def envs():
@@ -12,6 +19,18 @@ def envs():
 @pytest.fixture(scope= "session")
 def app_url():
 	return os.getenv("APP_URL")
+
+
+@pytest.fixture()
+def create_user(app_url):
+    new_user = {
+        "first_name": fake.first_name(),
+        "last_name": fake.last_name(),
+        "avatar": f"https://reqres.in/img/faces/{randint(1, 100)}-image.jpg",
+        "email": fake.free_email()}
+    user = requests.post(f"{app_url}/api/users", data=json.dumps(new_user))
+    assert user.status_code == HTTPStatus.CREATED
+    return user
 
 
 @pytest.fixture(scope="module")
